Updated Mutatation to use logged in user

+++ src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
+import getUserId from "../utils/getUserId";
 
 const Mutation = {
 	async createUser(parent, args, { prisma }, info) {
@@ -43,27 +44,27 @@ const Mutation = {
 			token: jwt.sign({ userId: user.id }, "thisisasecret"),
 		};
 	},
-	async deleteUser(parent, args, { prisma }, info) {
-		const userExists = await prisma.exists.User({ id: args.id });
+	async deleteUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
 
-		if (!userExists) {
-			throw new Error("User not found!");
-		}
-
-		return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+		return prisma.mutation.deleteUser({ where: { id: userId } }, info);
 	},
-	async updateUser(parent, args, { prisma }, info) {
+	async updateUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+
 		return prisma.mutation.updateUser(
 			{
 				where: {
-					id: args.id,
+					id: userId,
 				},
 				data: args.data,
 			},
 			info
 		);
 	},
-	createPost(parent, args, { prisma }, info) {
+	createPost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+
 		return prisma.mutation.createPost(
 			{
 				data: {
@@ -72,7 +73,7 @@ const Mutation = {
 					published: args.data.published,
 					author: {
 						connect: {
-							id: args.data.author,
+							id: userId,
 						},
 					},
 				},
@@ -80,7 +81,19 @@ const Mutation = {
 			info
 		);
 	},
-	deletePost(parent, args, { prisma }, info) {
+	async deletePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: {
+				id: userId,
+			},
+		});
+
+		if (!postExists) {
+			throw new Error("Unable to delete post");
+		}
+
 		return prisma.mutation.deletePost(
 			{
 				where: {
@@ -90,7 +103,19 @@ const Mutation = {
 			info
 		);
 	},
-	updatePost(parent, args, { prisma }, info) {
+	async updatePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: {
+				id: userId,
+			},
+		});
+
+		if (!postExists) {
+			throw new Error("Unable to update post");
+		}
+
 		return prisma.mutation.updatePost(
 			{
 				where: { id: args.id },
@@ -99,14 +124,16 @@ const Mutation = {
 			info
 		);
 	},
-	createComment(parent, args, { prisma }, info) {
+	createComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+
 		return prisma.mutation.createComment(
 			{
 				data: {
 					text: args.data.text,
 					author: {
 						connect: {
-							id: args.data.author,
+							id: userId,
 						},
 					},
 					post: {
@@ -119,7 +146,17 @@ const Mutation = {
 			info
 		);
 	},
-	deleteComment(parent, args, { prisma }, info) {
+	async deleteComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: {
+				id: userId,
+			},
+		});
+		if (!commentExists) {
+			throw new Error("Unable to delete Comment");
+		}
 		return prisma.mutation.deleteComment(
 			{
 				where: {
@@ -129,7 +166,17 @@ const Mutation = {
 			info
 		);
 	},
-	updateComment(parent, args, { prisma }, info) {
+	async updateComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request);
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: {
+				id: userId,
+			},
+		});
+		if (!commentExists) {
+			throw new Error("Unable to update Comment");
+		}
 		return prisma.mutation.updateComment(
 			{
 				where: {
